requirements.txt
whisperx==3.8.1
faster-whisper==1.2.1
torch==2.8.0
torchaudio==2.8.0
transformers==4.57.6
pyannote-audio==4.0.4
pyannote-core==6.0.1
pyannote-database==6.1.1
pyannote-metrics==4.0.0
pyannote-pipeline==4.0.0
ctranslate2==4.6.3
huggingface-hub==0.36.0
pandas==3.0.0
numpy==2.3.4
librosa==0.11.0
soundfile==0.13.1
pydub==0.25.1
python-dotenv==1.2.1
nltk==3.9.2
tqdm==4.67.2
ffmpeg-python==0.2.0
